Amend average_unique_users_daily to be a float and create waffle switch

+++ dataworkspace/dataworkspace/apps/datasets/models.py
@@ -273,7 +273,7 @@ class DataSet(DeletableTimestampedUserModel):
     search_vector = SearchVectorField(null=True, blank=True)
     subscriptions = GenericRelation(DataSetSubscription)
 
-    average_unique_users_daily = models.IntegerField(default=0)
+    average_unique_users_daily = models.FloatField(default=0)
 
     class Meta:
         db_table = "app_dataset"
@@ -1180,7 +1180,7 @@ class ReferenceDataset(DeletableTimestampedUserModel):
     search_vector = SearchVectorField(null=True, blank=True)
     subscriptions = GenericRelation(DataSetSubscription)
 
-    average_unique_users_daily = models.IntegerField(default=0)
+    average_unique_users_daily = models.FloatField(default=0)
 
     class Meta:
         db_table = "app_referencedataset"
@@ -2214,7 +2214,7 @@ class VisualisationCatalogueItem(DeletableTimestampedUserModel):
         null=True, blank=True, max_length=1024, help_text="Link to license (optional)"
     )
     search_vector = SearchVectorField(null=True, blank=True)
-    average_unique_users_daily = models.IntegerField(default=0)
+    average_unique_users_daily = models.FloatField(default=0)
 
     class Meta:
         permissions = [
